pyproject.toml
[project]
name = "g1-treasure-hunt"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[openai]",
    "httpx",
    "depthai",
    "opencv-python",
    "inference",
    "supervision",
    "numpy",
    "scipy",
]

[project.scripts]
agent = "main:main"

[dependency-groups]
dev = [
    "ruff",
    "basedpyright",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/g1"]

[tool.hatch.build.targets.wheel.force-include]
"src/main.py" = "main.py"

[tool.ruff]
target-version = "py311"
src = ["src", "devtools", "scripts"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific
    "FA",  # flake8-future-annotations
    "PTH", # flake8-use-pathlib
]

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
include = ["src", "devtools", "scripts"]
